requirements.txt
Flask==1.1.2
numpy==1.21.5
Pillow==9.4.0
scipy==1.7.3
